pyproject.toml
[tool.poetry]
name = "MLGO"
version = "0.1.0"
description = ""
authors = ["Tony Kabilan Okeke"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
notebook = "^6.4.10"
scikit-learn = "^1.2.1"
pandas = "^1.5.2"
scipy = "1.10"
hydra-core = "1.3"
infoml = "^0.8.0"
jupyter = "^1.0.0"
tqdm = "^4.64.1"
pyyaml = "^6.0"
statsmodels = "^0.13.5"

[tool.poetry.group.dev.dependencies]
interrogate = "^1.5.0"
black = "^22.12.0"
pdoc3 = "^0.10.0"
pytest = "^7.2.1"
pre-commit = "^3.0.2"
commitizen = "^2.41.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''
